requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
alembic==1.11.1
    # via
    #   -r requirements.in
    #   flask-migrate
aniso8601==9.0.1
    # via flask-restx
attrs==23.1.0
    # via jsonschema
blinker==1.6.2
    # via flask
build==0.10.0
    # via pip-tools
click==8.1.3
    # via
    #   flask
    #   pip-tools
colorama==0.4.6
    # via
    #   build
    #   click
flask==2.3.2
    # via
    #   -r requirements.in
    #   flask-jwt-extended
    #   flask-migrate
    #   flask-restx
    #   flask-sqlalchemy
flask-jwt-extended==4.5.2
    # via -r requirements.in
flask-migrate==4.0.4
    # via -r requirements.in
flask-restx==1.1.0
    # via -r requirements.in
flask-sqlalchemy==3.0.5
    # via
    #   -r requirements.in
    #   flask-migrate
greenlet==2.0.2
    # via sqlalchemy
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
jsonschema==4.17.3
    # via flask-restx
mako==1.2.4
    # via alembic
markupsafe==2.1.3
    # via
    #   jinja2
    #   mako
    #   werkzeug
packaging==23.1
    # via build
pip-tools==6.13.0
    # via -r requirements.in
psycopg2-binary==2.9.6
    # via -r requirements.in
pydantic==1.10.9
    # via -r requirements.in
pyjwt==2.7.0
    # via flask-jwt-extended
pyproject-hooks==1.0.0
    # via build
pyrsistent==0.19.3
    # via jsonschema
python-dotenv==1.0.0
    # via -r requirements.in
pytz==2023.3
    # via flask-restx
sqlalchemy==2.0.17
    # via
    #   alembic
    #   flask-sqlalchemy
tomli==2.0.1
    # via build
typing-extensions==4.6.3
    # via
    #   alembic
    #   pydantic
    #   sqlalchemy
werkzeug==2.3.6
    # via
    #   flask
    #   flask-jwt-extended
    #   flask-restx
wheel==0.40.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
